fixed issues that arose from merge
