+++ cyber-check/App.tsx
@@ -1,11 +1,8 @@
-import { StyleSheet,View,Text } from "react-native";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { StatusBar } from "expo-status-bar";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 
 import Signin from "./src/screens/SignIn";
-import Backend from "./src/screens/Backend";
+import RecentReportsScreen from "./src/screens/RecentReports";
 
 const Stack = createNativeStackNavigator();
 
@@ -22,52 +19,16 @@ const App = () => {
             }}
           />
           <Stack.Screen
-            name="Backend"
-            component={Backend}
+            name="RecentReportsTab"
+            component={RecentReportsScreen}
             options={{
               headerShown: false,
             }}
           />
         </Stack.Navigator>
       </NavigationContainer>
-    <NavigationContainer>
-      <View style={styles.container}>
-        <Text>Welcome to Cybercheck!</Text>
-        <StatusBar style="auto" />
-      </View>
-      <Tab.Navigator>
-        <Tab.Screen name="Home" component={HomeScreen} />
-        <Tab.Screen name="Recent Reports" component={RecentReportsScreen} />
-      </Tab.Navigator>
-    </NavigationContainer>
     </>
-
-  );
-}
-function HomeScreen() {
-  return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text>Home!</Text>
-    </View>
   );
 };
 
 export default App;
-function RecentReportsScreen() {
-  return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text>Recent Reports!</Text>
-    </View>
-  );
-}
-
-const Tab = createBottomTabNavigator();
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
