+++ cyber-check/package.json
@@ -9,24 +9,24 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@xmldom/xmldom": "^0.8.4",
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/bottom-tabs": "^6.4.0",
+    "@react-navigation/drawer": "^6.5.0",
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/native-stack": "^6.9.1",
     "@rneui/base": "^4.0.0-rc.7",
     "@rneui/themed": "^4.0.0-rc.7",
-    "@react-native-community/masked-view": "^0.1.11",
-    "@react-navigation/bottom-tabs": "^6.4.0",
-    "@react-navigation/drawer": "^6.5.0",
+    "@xmldom/xmldom": "^0.8.4",
     "expo": "~46.0.13",
     "expo-status-bar": "~1.4.0",
     "react": "18.0.0",
     "react-dom": "18.0.0",
     "react-native": "0.69.6",
+    "react-native-gesture-handler": "^2.8.0",
+    "react-native-reanimated": "^2.12.0",
     "react-native-safe-area-context": "4.3.1",
     "react-native-screens": "~3.15.0",
-    "react-native-web": "~0.18.7"
-    "react-native-gesture-handler": "^2.8.0",
-    "react-native-reanimated": "~2.9.1",
+    "react-native-web": "~0.18.7",
     "yarn": "^1.22.19"
   },
   "devDependencies": {
